+++ src/main/java/com/yanny/ytech/registration/YTechItemTags.java
@@ -52,7 +52,6 @@ public class YTechItemTags {
     public static final TagKey<Item> VENISON = create("venison");
     public static final TagKey<Item> WATER_BUCKETS = create("water_buckets");
 
-    public static final TagKey<Item> BONE = create("bone");
     public static final TagKey<Item> FERTILIZER = create("fertilizer");
     public static final TagKey<Item> FIRE_SOURCE = create("fire_source");
     public static final TagKey<Item> SOUL_FIRE_SOURCE = create("soul_fire_source");
@@ -112,29 +111,29 @@ public class YTechItemTags {
     public static final TypedTag<PartType> MOLDS = new PartTag("molds", PartType.ALL_PARTS);
 
     public static final TypedTag<MaterialType> ARROWS = new MaterialTag("arrows", ItemTags.ARROWS, YTechItems.ARROWS);
-    public static final TypedTag<MaterialType> AXES = new MaterialTag("axes", ItemTags.AXES, YTechItems.AXES);
-    public static final TypedTag<MaterialType> BOLTS = new MaterialTag("bolts", YTechItems.BOLTS);
-    public static final TypedTag<MaterialType> BOOTS = new MaterialTag("boots", Tags.Items.ARMORS_BOOTS, YTechItems.BOOTS);
-    public static final TypedTag<MaterialType> CHESTPLATES = new MaterialTag("chestplates", Tags.Items.ARMORS_CHESTPLATES, YTechItems.CHESTPLATES);
+    public static final TypedTag<MaterialType> AXES = new MaterialTag("axes", "forge", ItemTags.AXES, YTechItems.AXES);
+    public static final TypedTag<MaterialType> BOLTS = new MaterialTag("bolts", "forge", YTechItems.BOLTS);
+    public static final TypedTag<MaterialType> BOOTS = new MaterialTag("boots", "forge", Tags.Items.ARMORS_BOOTS, YTechItems.BOOTS);
+    public static final TypedTag<MaterialType> CHESTPLATES = new MaterialTag("chestplates", "forge", Tags.Items.ARMORS_CHESTPLATES, YTechItems.CHESTPLATES);
     public static final TypedTag<MaterialType> CRUSHED_MATERIALS = new MaterialTag("crushed_materials", "forge", YTechItems.CRUSHED_MATERIALS);
-    public static final TypedTag<MaterialType> FILES = new MaterialTag("files", YTechItems.FILES);
-    public static final TypedTag<MaterialType> HAMMERS = new MaterialTag("hammers", YTechItems.HAMMERS);
-    public static final TypedTag<MaterialType> HELMETS = new MaterialTag("helmets", Tags.Items.ARMORS_HELMETS, YTechItems.HELMETS);
-    public static final TypedTag<MaterialType> HOES = new MaterialTag("hoes", ItemTags.HOES, YTechItems.HOES);
+    public static final TypedTag<MaterialType> FILES = new MaterialTag("files", "forge", YTechItems.FILES);
+    public static final TypedTag<MaterialType> HAMMERS = new MaterialTag("hammers", "forge", YTechItems.HAMMERS);
+    public static final TypedTag<MaterialType> HELMETS = new MaterialTag("helmets", "forge", Tags.Items.ARMORS_HELMETS, YTechItems.HELMETS);
+    public static final TypedTag<MaterialType> HOES = new MaterialTag("hoes", "forge", ItemTags.HOES, YTechItems.HOES);
     public static final TypedTag<MaterialType> INGOTS = new IngotMaterialTag();
-    public static final TypedTag<MaterialType> KNIVES = new MaterialTag("knives", YTechItems.KNIVES);
-    public static final TypedTag<MaterialType> MORTAR_AND_PESTLES = new MaterialTag("mortar_and_pestles", YTechItems.MORTAR_AND_PESTLES);
-    public static final TypedTag<MaterialType> LEGGINGS = new MaterialTag("leggings", Tags.Items.ARMORS_LEGGINGS, YTechItems.LEGGINGS);
-    public static final TypedTag<MaterialType> PICKAXES = new MaterialTag("pickaxes", ItemTags.PICKAXES, YTechItems.PICKAXES);
-    public static final TypedTag<MaterialType> PLATES = new MaterialTag("plates", YTechItems.PLATES);
+    public static final TypedTag<MaterialType> KNIVES = new MaterialTag("knives", "forge", YTechItems.KNIVES);
+    public static final TypedTag<MaterialType> MORTAR_AND_PESTLES = new MaterialTag("mortar_and_pestles", "forge", YTechItems.MORTAR_AND_PESTLES);
+    public static final TypedTag<MaterialType> LEGGINGS = new MaterialTag("leggings", "forge", Tags.Items.ARMORS_LEGGINGS, YTechItems.LEGGINGS);
+    public static final TypedTag<MaterialType> PICKAXES = new MaterialTag("pickaxes", "forge", ItemTags.PICKAXES, YTechItems.PICKAXES);
+    public static final TypedTag<MaterialType> PLATES = new MaterialTag("plates", "forge", YTechItems.PLATES);
     public static final TypedTag<MaterialType> RAW_MATERIALS = new RawMaterialTag();
-    public static final TypedTag<MaterialType> RODS = new MaterialTag("rods", YTechItems.RODS);
-    public static final TypedTag<MaterialType> SAWS = new MaterialTag("saws", YTechItems.SAWS);
-    public static final TypedTag<MaterialType> SHEARS = new MaterialTag("shears", "forge", YTechItems.SHEARS);
+    public static final TypedTag<MaterialType> RODS = new MaterialTag("rods", "forge", Tags.Items.RODS, YTechItems.RODS);
+    public static final TypedTag<MaterialType> SAWS = new MaterialTag("saws", "forge", YTechItems.SAWS);
+    public static final TypedTag<MaterialType> SHEARS = new MaterialTag("shears", "forge", Tags.Items.SHEARS, YTechItems.SHEARS);
     public static final TypedTag<MaterialType> SAW_BLADES = new MaterialTag("saw_blades", YTechItems.SAW_BLADES);
-    public static final TypedTag<MaterialType> SHOVELS = new MaterialTag("shovels", ItemTags.SHOVELS, YTechItems.SHOVELS);
+    public static final TypedTag<MaterialType> SHOVELS = new MaterialTag("shovels", "forge", ItemTags.SHOVELS, YTechItems.SHOVELS);
     public static final TypedTag<MaterialType> SPEARS = new MaterialTag("spears", YTechItems.SPEARS);
-    public static final TypedTag<MaterialType> SWORDS = new MaterialTag("swords", ItemTags.SWORDS, YTechItems.SWORDS);
+    public static final TypedTag<MaterialType> SWORDS = new MaterialTag("swords", "forge", ItemTags.SWORDS, YTechItems.SWORDS);
 
     public static final TypedTag<MaterialType> DEEPSLATE_ORES = new DeepslateOreMaterialTag();
     public static final TypedTag<MaterialType> DRYING_RACKS = new MaterialTag(YTechBlockTags.DRYING_RACKS);